refactor: Make UserConversation and Message entities support staff members
The changes modify the `UserConversation` and `Message` entities to better support staff members in the conversation system.

In `UserConversation`, the `UserId` is made nullable and a new `StaffId` property with corresponding `Staff` navigation property is added.

In `Message`, the `Sender` property is commented out and replaced with an `IsStaff` flag to distinguish between user and staff senders. These changes enable the system to handle conversations involving both users and staff members.