refactor: Change `IsActivated` type from bool to string in `UpdateClinicCommand`
The `IsActivated` property in `UpdateClinicCommand` was modified from `bool?` to `string?` to accommodate more flexible activation states that may need to be represented as strings rather than simple boolean values. This change allows for handling different activation status formats in the API. No breaking changes are introduced as the property remains nullable.